Moved the exceptions to their own class and added the NoCollectionsURL exception
